% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add_bin_begin_and_end}
\alias{add_bin_begin_and_end}
\title{Add begin and end for a binned matrix}
\usage{
add_bin_begin_and_end(mat)
}
\description{
Add begin and end for a binned matrix
}
\keyword{internal}
