% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{add_bin_bounds_and_distance}
\alias{add_bin_bounds_and_distance}
\title{Add begin and end for a binned matrix, as well as middle position and average distance}
\usage{
add_bin_bounds_and_distance(mat)
}
\description{
Add begin and end for a binned matrix, as well as middle position and average distance
}
\examples{
" "
}
\keyword{internal}
