% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppDocs.R
\name{bin_cts_to_mat}
\alias{bin_cts_to_mat}
\title{Build binned matrix using raw cts data}
\usage{
bin_cts_to_mat(cs, cts, resolution)
}
\arguments{
\item{cs}{}

\item{cts}{}

\item{resolution}{}
}
\description{
Build binned matrix using raw cts data
}
