% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{bin_data}
\alias{bin_data}
\title{Bin a reads data.table into a matrix of a given resolution}
\usage{
bin_data(obj, resolution, b1 = NULL, b2 = NULL)
}
\arguments{
\item{obj}{a CSnorm object, or a data.table containing data to be binned}

\item{resolution}{the target resolution}

\item{b2, e2}{(optional) if provided, give extradiagonal portion}
}
\value{
a data.table representing the binned data
}
\description{
Bin a reads data.table into a matrix of a given resolution
}
