% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_graph.R
\name{build_patch_graph}
\alias{build_patch_graph}
\title{give a patch ID to each patch in a binless matrix, and report local extrema}
\usage{
build_patch_graph(mat, g, tol.value = 0.001)
}
\description{
give a patch ID to each patch in a binless matrix, and report local extrema
}
\keyword{internal}
