% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{categorize_by_new_type}
\alias{categorize_by_new_type}
\title{Add category label to reads}
\usage{
categorize_by_new_type(sub, maxlen = 600, dangling.L = c(0),
  dangling.R = c(3), read.len = 40)
}
\arguments{
\item{sub}{reads data.table to cateogorize}

\item{maxlen}{positive integer.  Maximum admissible size of sonication 
fragment. Fragments above this threshold will be categorized as "other".}

\item{dangling.L, dangling.R}{vectors of integers. Offset from cut site for
reads to be considered dangling ends (Left or Right respectively)}

\item{read.len}{integer vector. The length of a full read, used for dangling ends (temporary)}
}
\value{
The same data.table, with an additional "category" column
}
\description{
Performs classification according to cut-site normalization scheme
}
\examples{
" "
}
