% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{check_fit}
\alias{check_fit}
\title{Verify model fit by computing posterior predictive quantities}
\usage{
check_fit(cs, genomic.groups = 5, decay.groups = 5, npoints = 10)
}
\arguments{
\item{cs}{CSnorm object, normalized.}

\item{genomic.groups}{How many groups for iota and rho}

\item{decay.groups}{How many groups for diagonal decay}

\item{npoints}{Number of points per group to take (not guaranteed)}
}
\description{
Verify model fit by computing posterior predictive quantities
}
