% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{compress_cts}
\alias{compress_cts}
\title{make a summarized cts matrix used for one IRLS iteration}
\usage{
compress_cts(cs, cts)
}
\description{
make a summarized cts matrix used for one IRLS iteration
}
\keyword{internal}
