% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_graph.R
\name{compute_2d_connectivity_graph}
\alias{compute_2d_connectivity_graph}
\title{connectivity on a triangle}
\usage{
compute_2d_connectivity_graph(nbins, start = 1)
}
\description{
connectivity on a triangle
}
\keyword{internal}
