% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppDocs.R
\name{create_empty_matrix}
\alias{create_empty_matrix}
\title{Generate an empty matrix in triangular form.}
\usage{
create_empty_matrix(name, bins)
}
\arguments{
\item{name}{an ordered factor of names}

\item{bins}{an ordered factor of bins}
}
\value{

}
\description{
arguments are taken as ordered factors
returned matrix is of size n_names * n_bins * (n_bins+1) / 2
n_names and n_bins are the number of levels in the name and bins factors
}
