% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binless.R
\docType{class}
\name{CSdata-class}
\alias{CSdata-class}
\title{Class to hold a single experiment}
\description{
Class to hold a single experiment
}
\section{Slots}{

\describe{
\item{\code{info}}{list. Information on this experiment}

\item{\code{settings}}{list. Settings for the normalization procedures}

\item{\code{data}}{data.table. The reads data for that experiment}

\item{\code{biases}}{data.table. The cut-site information for that experiment}

\item{\code{counts}}{data.table. The reads data for that experiment}
}}

