% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binless.R
\docType{class}
\name{CSnorm-class}
\alias{CSnorm-class}
\title{Class to hold cut-site normalization data}
\description{
Class to hold cut-site normalization data
}
\section{Slots}{

\describe{
\item{\code{experiments}}{data.table. Essential information for each experiment}

\item{\code{design}}{data.table. A design matrix for the normalization}

\item{\code{settings}}{list. Additional parameters for the normalization}

\item{\code{biases}}{data.table. Merged bias information for all experiments}

\item{\code{counts}}{data.table. Merged count information for all experiments}

\item{\code{par}}{list. Parameters of normalization}

\item{\code{diagnostics}}{list. Diagnostics output of parallel runs}

\item{\code{pred}}{list. Predicted quantities}

\item{\code{binned}}{list. Binned matrices and differences}
}}

