% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{detect_binless_differences}
\alias{detect_binless_differences}
\title{Binless detection of significant differences with a reference}
\usage{
detect_binless_differences(cs, ref, resolution = cs@settings$base.res,
  group = "all", ncores = 1, tol.val = cs@settings$tol, verbose = T,
  fix.lambda1 = F, fix.lambda1.at = NA, fix.lambda2 = F,
  fix.lambda2.at = NA, nperf = 50)
}
\arguments{
\item{cs}{}

\item{ref}{}

\item{resolution}{}

\item{group}{}

\item{ncores}{}
}
\value{
the binned matrix with additional information relating to these
  significant interactions
}
\description{
Binless detection of significant differences with a reference
}
