% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_graph.R
\name{detect_binless_patches}
\alias{detect_binless_patches}
\title{give a patch ID to each patch in a binless matrix, and report local extrema}
\usage{
detect_binless_patches(mat, settings)
}
\description{
give a patch ID to each patch in a binless matrix, and report local extrema
}
\keyword{internal}
