% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binned.R
\name{detect_binned_interactions}
\alias{detect_binned_interactions}
\title{Binned detection of significant interactions wrt expected}
\usage{
detect_binned_interactions(cs, resolution = cs@settings$base.res,
  group = "all", threshold = 0.95, ncores = 1, niter = 100,
  tol = cs@settings$tol, verbose = T)
}
\arguments{
\item{cs}{CSnorm object}

\item{resolution, group}{see
\code{\link{bin_all_datasets}} and \code{\link{group_datasets}}, used to
identify the input matrices.}

\item{threshold}{on the probability K/(1+K) where K is the Bayes factor}

\item{ncores}{number of cores used for parallelization}

\item{niter}{integer. Maximum number of IRLS iterations}

\item{tol}{numeric. Convergence tolerance for IRLS objective}

\item{verbose}{boolean.}
}
\value{
the binned matrix with additional information relating to these 
  significant interactions
}
\description{
Binned detection of significant interactions wrt expected
}
\examples{
" "
}
