% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{di_index}
\alias{di_index}
\title{Compute directionality index on a matrix}
\usage{
di_index(hicmat, nw, n, triag = T)
}
\arguments{
\item{hicmat, nw, n, triag}{}
}
\description{
Compute directionality index on a matrix
}
\keyword{internal}
