% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{examine_dataset}
\alias{examine_dataset}
\title{Diagnostic plots to determine dangling.L, dangling.R and maxlen arguments}
\usage{
examine_dataset(infile, window = 15, maxlen = 1000, skip.fbm = T,
  read.len = 40, skip = 0L, nrows = -1L, locus = NULL)
}
\arguments{
\item{infile}{A tadbit .tsv file as character string, or a data.table returned by read.tsv}

\item{window}{how many bases to plot around cut site}

\item{maxlen}{how many bases to plot off-diagonal}

\item{skip.fbm}{boolean. If TRUE (default), skip fragment-based IDs, 
containing # and ~. Otherwise keep them.}

\item{read.len}{length of reads admissible for dangling end candidates}

\item{skip}{see \code{\link[data.table]{fread}}}

\item{nrows}{see \code{\link[data.table]{fread}}}

\item{locus}{if not NULL, a vector in the format c("chr3",123,456) to keep only
paired-end reads where both ends map in region chr3:123-456. Otherwise, read everything.}
}
\value{
Two plots, pdangling to determine dangling.L and dangling.R, and
  pdiag for maxlen. Also returns data, the reads data table used for the
  plots.
}
\description{
Diagnostic plots to determine dangling.L, dangling.R and maxlen arguments
}
\examples{
" "
}
