% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{fit}
\alias{fit}
\title{Single-cpu fitting}
\usage{
fit(biases, counts, design, dmin, dmax, bf_per_kb = 1, bf_per_decade = 20,
  iter = 10000, verbose = T, init = 0, weight = array(1, dim = design[,
  .N]), ...)
}
\description{
Single-cpu fitting
}
\keyword{internal}
