% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{fit_fixed}
\alias{fit_fixed}
\title{Single-cpu fitting, fixed length scales (custom for genomic, lambda=1 for decay)}
\usage{
fit_fixed(biases, counts, design, dmin, dmax, lambda, bf_per_kb = 1,
  bf_per_decade = 20, iter = 10000, verbose = T, weight = array(1, dim =
  design[, .N]), ...)
}
\description{
Single-cpu fitting, fixed length scales (custom for genomic, lambda=1 for decay)
}
\keyword{internal}
