% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{fuse_close_cut_sites}
\alias{fuse_close_cut_sites}
\title{Fuse cut sites that are very close to each other}
\usage{
fuse_close_cut_sites(biases, counts, dfuse, name, circularize)
}
\value{
a list with filtered biases and counts
}
\description{
Fuse cut sites that are very close to each other
}
\keyword{internal}
