% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{gauss_common_muhat_mean}
\alias{gauss_common_muhat_mean}
\title{Compute means for positive and zero counts using previous params}
\usage{
gauss_common_muhat_mean(cs, zeros, sbins)
}
\value{
data.table containing the following columns
- name: name of the dataset
- id1, pos1, bin1: coordinates of the cut site
- bin2, dbin: signal/distance bin in which we looked at the intersections
- dir: fwd (rev) for contacts with downstream (resp upstream) cut-sites
- cat: whether we consider contacts on the left (contact L) or on the right (contact R) of this cut site
- count: the value of the count in this signal/distance/direction/category bin.
 We discard anything below cs@settings$dmin. Note that sum(count) is twice the total number of counts (per dataset).
- lmu.nosig: the log(mean) of the background model (exposure+bias+decay)
- phi: the log(signal)
- mu: the mean, including signal
- nobs: how many observations in this signal/distance/direction/category bin have this count.
 Note that sum(nobs) is twice the total number of observable counts (per dataset).
- eC, log_decay, log_bias: exposure, log(decay) and log(bias) of this dataset. The bias is either iota or rho depending on cat.
- z: count/mu-1
- var: 1/mu+1/dispersion
}
\description{
Compute means for positive and zero counts using previous params
}
\keyword{internal}
