% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{gauss_common_muhat_mean_biases}
\alias{gauss_common_muhat_mean_biases}
\title{Compute predicted means for dangling and rejoined}
\usage{
gauss_common_muhat_mean_biases(cs)
}
\description{
Compute predicted means for dangling and rejoined
}
\keyword{internal}
