% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{gauss_decay}
\alias{gauss_decay}
\title{Single-cpu simplified fitting for iota and rho}
\usage{
gauss_decay(cs, cts.common, verbose = T, update.eC = T)
}
\description{
Single-cpu simplified fitting for iota and rho
}
\keyword{internal}
