% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_decay.R
\name{gauss_decay_muhat_mean}
\alias{gauss_decay_muhat_mean}
\title{Compute decay etahat and weights using previous mean}
\usage{
gauss_decay_muhat_mean(cs, cts.common)
}
\description{
Compute decay etahat and weights using previous mean
}
\keyword{internal}
