% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{gauss_decay_optimize}
\alias{gauss_decay_optimize}
\title{Optimize decay parameters}
\usage{
gauss_decay_optimize(csd, design, Kdiag, original_lambda_diag, verbose = T,
  max_perf_iteration = 1000, convergence_epsilon = 1e-09)
}
\description{
Optimize decay parameters
}
\keyword{internal}
