% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_dispersion.R
\name{gauss_dispersion}
\alias{gauss_dispersion}
\title{Single-cpu simplified fitting for exposures and dispersion}
\usage{
gauss_dispersion(cs, cts.common, verbose = T)
}
\description{
Single-cpu simplified fitting for exposures and dispersion
}
\keyword{internal}
