% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_exposures.R
\name{gauss_exposures}
\alias{gauss_exposures}
\title{Compute eC using previous mean}
\usage{
gauss_exposures(cs, cts.common, verbose = T)
}
\description{
Compute eC using previous mean
}
\keyword{internal}
