% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_signal.R
\name{gauss_signal}
\alias{gauss_signal}
\title{fit signal using sparse fused lasso}
\usage{
gauss_signal(cs, cts.common, verbose = T, ncores = 1, fix.lambda1 = F,
  fix.lambda1.at = NA, fix.lambda2 = F, fix.lambda2.at = NA)
}
\description{
fit signal using sparse fused lasso
}
\keyword{internal}
