% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{gauss_signal_muhat_mean}
\alias{gauss_signal_muhat_mean}
\title{fit signal using sparse fused lasso}
\usage{
gauss_signal_muhat_mean(cs, cts.common)
}
\description{
fit signal using sparse fused lasso
}
\keyword{internal}
