% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{generate_cubic_spline}
\alias{generate_cubic_spline}
\title{Generate cubic spline}
\usage{
generate_cubic_spline(cutsites, Krow, sparse = F)
}
\description{
Generate cubic spline
}
\keyword{internal}
