% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{generate_fake_dataset}
\alias{generate_fake_dataset}
\title{Generate a fake dataset that follows the cut site normalization model's specifications}
\usage{
generate_fake_dataset(biases.ref = NULL, num_rsites = 3000,
  genome_size = 1e+06, eC = -4.4, eRJ = 3, eDE = 1, alpha = 2,
  condition = "WT", replicate = "1", enzyme = "NA",
  name = paste("Fake", condition, enzyme, replicate), dmin = 1000,
  signal = F)
}
\arguments{
\item{num_rsites}{Number of restriction sites}

\item{genome_size}{Size of the genome}

\item{eC, eRJ, eDE}{Exposure for counts, rejoined and dangling ends}

\item{alpha}{Dispersion}
}
\value{
a list of biases and counts, similar to \code{\link{prepare_for_sparse_cs_norm}}
}
\description{
Generate a fake dataset that follows the cut site normalization model's specifications
}
