% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{generate_genomic_biases}
\alias{generate_genomic_biases}
\title{Generate iota and rho genomic biases on evenly spaced points along the genome}
\usage{
generate_genomic_biases(biases, beta_iota, beta_rho, bf_per_kb = 1,
  points_per_kb = 100)
}
\arguments{
\item{biases}{data.table.}

\item{beta_iota, beta_rho}{vectors. spline parameters}

\item{bf_per_kb}{number of basis functions per kb}

\item{points_per_kb}{number of evaluation points per kb}
}
\description{
Generate iota and rho genomic biases on evenly spaced points along the genome
}
\keyword{internal}
