% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_binless_differences}
\alias{get_binless_differences}
\title{Fetch binless differences from CSnorm object}
\usage{
get_binless_differences(cs, ref, resolution = cs@settings$base.res,
  group = "all")
}
\arguments{
\item{cs}{CSnorm object.}

\item{resolution}{The resolution of the matrix, in bases}

\item{group}{character. Either "all" to return individual matrices, or any of 
"condition", "replicate" or "enzyme" to return grouped matrices}
}
\value{
a data.table containing the differences
}
\description{
Fetch binless differences from CSnorm object
}
