% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_binless_interactions}
\alias{get_binless_interactions}
\title{Fetch binless interactions from CSnorm object}
\usage{
get_binless_interactions(cs, resolution = cs@settings$base.res,
  group = "all")
}
\arguments{
\item{cs}{CSnorm object.}

\item{resolution}{The resolution of the matrix, in bases}

\item{group}{character. Either "all" to return individual matrices, or any of 
"condition", "replicate" or "enzyme" to return grouped matrices}
}
\value{
a data.table containing the interactions
}
\description{
Fetch binless interactions from CSnorm object
}
\examples{
" "
}
