% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_binned_matrices}
\alias{get_binned_matrices}
\title{Fetch grouped matrices from CSnorm object}
\usage{
get_binned_matrices(cs, resolution = cs@settings$base.res,
  group = "all")
}
\arguments{
\item{cs}{CSnorm object.}

\item{resolution}{The resolution of the matrix, in bases}

\item{group}{character. Either "all" to return individual matrices, or any of 
"condition", "replicate" or "enzyme" to return grouped matrices}
}
\value{
a data.table containing the binned matrices
}
\description{
Fetch grouped matrices from CSnorm object
}
\examples{
" "
}
