% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_cs_group_idx}
\alias{get_cs_group_idx}
\title{Fetch CSgroup indices from CSnorm object}
\usage{
get_cs_group_idx(cs, resolution, group, raise = T)
}
\arguments{
\item{cs}{CSnorm object}

\item{resolution}{}

\item{raise}{boolean. If T raise an exception, otherwise return -1.}
}
\value{
CSgroup object index
}
\description{
Fetch CSgroup indices from CSnorm object
}
\keyword{internal}
