% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_cs_interaction_idx}
\alias{get_cs_interaction_idx}
\title{Fetch CSinter object from CSgroup object}
\usage{
get_cs_interaction_idx(csg, type, threshold = -1, ref = NULL,
  raise = T)
}
\arguments{
\item{csg}{csg}

\item{type}{type}

\item{threshold}{threshold}

\item{ref}{ref}

\item{raise}{raise}
}
\value{
" "
}
\description{
Fetch CSinter object from CSgroup object
}
\examples{
" "
}
\keyword{internal}
