% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_exact_logp}
\alias{get_exact_logp}
\title{Return the value of the log posterior of the exact model (memory-intensive)}
\usage{
get_exact_logp(cs)
}
\arguments{
\item{cs}{an optimized csnorm object}
}
\value{
The log-posterior
}
\description{
Return the value of the log posterior of the exact model (memory-intensive)
}
