% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_genomic_biases}
\alias{get_genomic_biases}
\title{Predict genomic biases at evenly spaced positions across the genome}
\usage{
get_genomic_biases(cs, points_per_kb = 10)
}
\arguments{
\item{cs}{}

\item{points_per_kb}{}
}
\description{
Predict genomic biases at evenly spaced positions across the genome
}
