% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_interactions}
\alias{get_interactions}
\title{Fetch detected interactions from CSnorm object}
\usage{
get_interactions(cs, type, resolution, group, threshold = -1,
  ref = NULL)
}
\description{
Fetch detected interactions from CSnorm object
}
\examples{
" "
}
\keyword{internal}
