% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_ligation_ratio}
\alias{get_ligation_ratio}
\title{Compute the ligation ratio from a TadBit unfiltered tsv file}
\usage{
get_ligation_ratio(fname, maxlen, read.len, dangling.L, dangling.R,
  locus = NULL)
}
\arguments{
\item{fname}{a tsv file path}

\item{maxlen}{positive integer.  Maximum admissible size of sonication 
fragment. Fragments above this threshold will be categorized as "other".}

\item{read.len}{integer vector. The length of a full read, used for dangling ends (temporary)}

\item{dangling.L}{vectors of integers. Offset from cut site for
reads to be considered dangling ends (Left or Right respectively)}

\item{dangling.R}{vectors of integers. Offset from cut site for
reads to be considered dangling ends (Left or Right respectively)}

\item{locus}{if not NULL, a vector in the format c("chr3",123,456) to keep only
paired-end reads where both ends map in region chr3:123-456. Otherwise, read everything.}
}
\value{
The ligation ratio and the number of dangling and other ends
}
\description{
Compute the ligation ratio from a TadBit unfiltered tsv file
}
