% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_predicted_subset}
\alias{get_predicted_subset}
\title{Predict model parameters on a subset of the input data}
\usage{
get_predicted_subset(cs, ncounts = 1e+05)
}
\arguments{
\item{cs}{}

\item{ncounts}{}
}
\description{
Predict model parameters on a subset of the input data
}
