% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_raw_reads}
\alias{get_raw_reads}
\title{Get subset of raw reads}
\usage{
get_raw_reads(data, b1, e1, b2 = NULL, e2 = NULL)
}
\arguments{
\item{data}{reads data.table containing the raw reads}

\item{b1, e1}{begin / end of read range you want}

\item{b2, e2}{(optional) if provided, give extradiagonal portion}
}
\value{
reads data.table subset
}
\description{
Get subset of raw reads
}
