% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{get_residuals}
\alias{get_residuals}
\title{Check whether a normalization has converged}
\usage{
get_residuals(cts.common, viewpoint)
}
\description{
Check whether a normalization has converged
}
\keyword{internal}
