% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{get_signal_matrix}
\alias{get_signal_matrix}
\title{Build (grouped) signal matrix using normalization data if available}
\usage{
get_signal_matrix(cs, resolution = cs@settings$base.res,
  groups = cs@experiments[, .(name, groupname = name)])
}
\description{
Build (grouped) signal matrix using normalization data if available
}
\keyword{internal}
