% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{gfl_BIC}
\alias{gfl_BIC}
\title{compute BIC for a given value of lambda2, optimizing lambda1 and eCprime (performance iteration, persistent state)}
\usage{
gfl_BIC(csig, lambda2, constrained = T, positive = T, fixed = F,
  fix.lambda1 = F, fix.lambda1.at = NA)
}
\description{
compute BIC for a given value of lambda2, optimizing lambda1 and eCprime (performance iteration, persistent state)
}
\keyword{internal}
