% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{gfl_compute_initial_state}
\alias{gfl_compute_initial_state}
\title{build initial state from phi / delta}
\usage{
gfl_compute_initial_state(csig, diff = F)
}
\description{
build initial state from phi / delta
}
\keyword{internal}
