% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{group_datasets}
\alias{group_datasets}
\title{Group binned matrices of datasets}
\usage{
group_datasets(cs, resolution, group = c("condition", "replicate",
  "enzyme", "experiment"), verbose = T, ncores = 1)
}
\arguments{
\item{cs}{CSnorm object, normalized.}

\item{resolution}{integer. The desired resolution of the matrix.}

\item{group}{The type of grouping to be performed. Any combination of the given arguments is possible.}

\item{verbose}{verbose}

\item{ncores}{integer. The number of cores to parallelize the zeros calculation on.}
}
\value{
CSnorm object
}
\description{
Group binned matrices of datasets
}
\examples{
" "
}
