% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{guarantee_beta_diag_increasing}
\alias{guarantee_beta_diag_increasing}
\title{Ensure beta_diag is strictly increasing to make stan happy}
\usage{
guarantee_beta_diag_increasing(beta_diag)
}
\description{
Ensure beta_diag is strictly increasing to make stan happy
}
\keyword{internal}
