% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{initial_guess_exposures}
\alias{initial_guess_exposures}
\title{Compute initial exposures assuming a poisson model}
\usage{
initial_guess_exposures(cs, cts.common, pseudocount = 0.01)
}
\description{
Compute initial exposures assuming a poisson model
}
\keyword{internal}
