% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{initial_guess_genomic}
\alias{initial_guess_genomic}
\title{Compute initial iota and rho biases assuming a poisson model}
\usage{
initial_guess_genomic(cs, cts.common, pseudocount = 0.01)
}
\description{
Compute initial iota and rho biases assuming a poisson model
}
\keyword{internal}
