% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ice.R
\name{iterative_normalization}
\alias{iterative_normalization}
\title{Apply the ICE algorithm to a binned matrix}
\usage{
iterative_normalization(mat, niterations = 100, namecol = "name",
  verbose = T)
}
\arguments{
\item{mat}{a matrix obtained by grouping}

\item{niterations}{positive integer. Number of iterations to perform}
}
\value{
a CSbinned object containing the ICEd matrix
}
\description{
Apply the ICE algorithm to a binned matrix
}
\examples{
" "
}
