% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{load_stripped}
\alias{load_stripped}
\title{Load normalized CS object saved with save_stripped,
rebuild zeros data.table}
\usage{
load_stripped(fname, ncores = 1)
}
\arguments{
\item{fname}{file name}

\item{ncores}{number of CPUs to use to reconstruct zeros}
}
\value{
A cs object on which signal detection can be performed
}
\description{
Load normalized CS object saved with save_stripped,
rebuild zeros data.table
}
\examples{
" "
}
