% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{merge_cs_norm_datasets}
\alias{merge_cs_norm_datasets}
\title{Merge one or more CSdata objects into a CSnorm object and set experimental design}
\usage{
merge_cs_norm_datasets(datasets, different.decays = c("none", "all", "enzyme",
  "condition"), dfuse = 5, qmin = 0.05)
}
\arguments{
\item{datasets}{list of CSdata objects}

\item{different.decays}{character. If "none" (default), one decay is modelled
for all experiments. If "all", each experiment has its own decay. If
"enzyme", experiments with a different enzyme get their own decay. If
"condition", experiments with a different condition get their own decay.
Arguments can be abbreviated and combined}

\item{dfuse}{Fuse cut sites that are closer than dfuse to each other}
}
\value{
CSnorm object
}
\description{
This will do some cleanup on the data, namely:
- remove all contacts lower than their respective dmin
- remove the counts at cut sites with too low or too high coverage (default 1st and 99th percentile)
Note that you can customize the design, as long as you keep the numbering sequential
and starting at 1.
}
