% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_binless}
\alias{normalize_binless}
\title{Binless normalization}
\usage{
normalize_binless(cs, restart = F, bf_per_kb = 50, bf_per_decade = 10,
  bins_per_bf = 10, base.res = 5000, ngibbs = 15, bg.steps = 5,
  iter = 100, fit.signal = T, verbose = T, init.dispersion = 1,
  tol = 0.1, ncores = 1, fix.lambda1 = F, fix.lambda1.at = NA,
  fix.lambda2 = F, fix.lambda2.at = NA)
}
\arguments{
\item{cs}{CSnorm object as returned by \code{\link{merge_cs_norm_datasets}}}

\item{bf_per_kb}{positive numeric. Number of cubic spline basis functions per
kilobase, for genomic bias estimates. Small values make the optimization 
easy, but makes the genomic biases stiffer.}

\item{bf_per_decade}{positive numeric. Number of cubic spline basis functions
per distance decade (in bases), for diagonal decay. Default parameter 
should suffice.}

\item{bins_per_bf}{positive integer. Number of distance bins to split basis 
functions into. Must be sufficiently small so that the diagonal decay is 
approximately constant in that bin.}

\item{ngibbs}{positive integer. Number of gibbs sampling iterations.}

\item{iter}{positive integer. Maximum number of optimization steps for each leg.}

\item{fit.signal}{boolean. Set to FALSE only for diagnostics.}

\item{verbose}{Display progress if TRUE}

\item{init.dispersion}{positive numeric. Value of the dispersion to use initially.}

\item{tol}{positive numeric (default 1e-2). Convergence tolerance on relative changes in the computed biases.}

\item{ncores}{positive integer (default 1). Number of cores to use.}

\item{fix.lambda1}{whether to set lambda1 to a given value, or to estimate it}

\item{fix.lambda1.at}{if fix.lambda1==T, the approximate value where it is meant to be fixed. Might move a bit because
of the positivity and degeneracy constraints.}

\item{init}{boolean (default is FALSE). Whether to continue an existing
normalization or to start anew (default).}

\item{lambdas}{positive numeric. Length scales to try out as initial
condition.}
}
\value{
A csnorm object
}
\description{
Binless normalization
}
\examples{

}
