% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso.R
\name{optimize_lambda2}
\alias{optimize_lambda2}
\title{cross-validate lambda2 assuming it is very rugged}
\usage{
optimize_lambda2(csig, n.SD = 1, constrained = T, positive = T,
  fixed = F, fix.lambda1 = F, fix.lambda1.at = NA, ncores = 1)
}
\description{
cross-validate lambda2 assuming it is very rugged
}
\keyword{internal}
