% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso.R
\name{optimize_lambda2_smooth}
\alias{optimize_lambda2_smooth}
\title{cross-validate lambda2 assuming it is smooth}
\usage{
optimize_lambda2_smooth(csig, constrained = T, positive = T,
  fixed = F, min.lambda2 = 0.1, fix.lambda1 = F,
  fix.lambda1.at = NA)
}
\description{
cross-validate lambda2 assuming it is smooth
}
\keyword{internal}
