% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{optimize_stan_model}
\alias{optimize_stan_model}
\title{Initial guess for normalization}
\usage{
optimize_stan_model(model, data, iter, verbose, init, ...)
}
\description{
Initial guess for normalization
}
\keyword{internal}
