% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot_binless_matrix}
\alias{plot_binless_matrix}
\title{make plot of binless matrix}
\usage{
plot_binless_matrix(mat, upper = "binless", lower = "observed",
  trans = "log10", facet = "name", limits = NULL, label = NA)
}
\arguments{
\item{mat}{the binless matrix}

\item{upper, lower}{the name of the column to plot in upper (resp. lower)
triangle. Default: phi. Formulae are allowed (ggplot2 aes syntax)}

\item{facet}{what to facet on. Default: name. Pass a character vector for multiple facets}

\item{limits}{set to a pair of values to enforce plot and colour scale to be within these values}
}
\description{
make plot of binless matrix
}
