% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot_binned}
\alias{plot_binned}
\title{Plot reads in binned form at a given resolution}
\usage{
plot_binned(dt, resolution, b1, e1, b2 = NULL, e2 = NULL,
  diagonal = F, rsites = F)
}
\arguments{
\item{dt}{a reads data.table containing the parsed tsv reads}

\item{resolution}{the requested resolution, in base pairs}

\item{b1, e1}{the begin and end of the range requested}

\item{b2, e2}{optional, if you want an outer-diagonal section}

\item{diagonal}{logical, default FALSE. Should the diagonal be indicated?}

\item{rsites}{logical, default FALSE. Should restriction sites be indicated?}
}
\description{
Plot reads in binned form at a given resolution
}
