% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot_diagnostics}
\alias{plot_diagnostics}
\title{Diagnostics plots to monitor convergence of normalization (gaussian
approximation)}
\usage{
plot_diagnostics(cs, start = 1, end = Inf)
}
\arguments{
\item{cs}{a normalized CSnorm object}

\item{start, end}{the starting and ending steps on which to plot the info. Default is full range.}
}
\value{
Two plots in a list. The first is for the four log-likelihoods, the
  second is for the parameters.
}
\description{
Diagnostics plots to monitor convergence of normalization (gaussian
approximation)
}
