% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{plot_raw}
\alias{plot_raw}
\title{Plot raw reads}
\usage{
plot_raw(dt, b1 = NULL, e1 = NULL, b2 = NULL, e2 = NULL, diagonal = T,
  rsites = T)
}
\arguments{
\item{dt}{a reads data.table containing the parsed tsv reads}

\item{b1}{the begin and end of the range requested}

\item{e1}{the begin and end of the range requested}

\item{b2}{optional, if you want an outer-diagonal section}

\item{e2}{optional, if you want an outer-diagonal section}

\item{diagonal}{logical, default FALSE. Should the diagonal be indicated?}

\item{rsites}{logical, default FALSE. Should restriction sites be indicated?}
}
\description{
Plot raw reads
}
