% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{plot_virtual_4C}
\alias{plot_virtual_4C}
\title{Make a virtual 4C plot}
\usage{
plot_virtual_4C(mat, start, width = 5000)
}
\arguments{
\item{mat}{the binless matrix data.table}

\item{start, width}{the start and width of the 4C viewpoint. Will be rounded to multiples
of the matrix base resolution. If width spans several base.res intervals, will average
the resulting observations based on the distance of each interaction}
}
\value{
" "
}
\description{
Will show value of binless matrix at this viewpoint, along with (rescaled) original observations
}
\examples{
" "
}
