% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{predict_all}
\alias{predict_all}
\title{Predict expected values for each count given optimized model parameters}
\usage{
predict_all(cs, counts, verbose = T)
}
\arguments{
\item{cs}{CSnorm object}

\item{verbose}{}
}
\value{
a stan optimization output with the predictions.
}
\description{
Predict expected values for each count given optimized model parameters
}
\section{Warning}{
 Do not call this function to compute a binned matrix. It
  should (almost) never be used and might be removed in the future.
}

\keyword{internal}
