% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{predict_all_parallel}
\alias{predict_all_parallel}
\title{Fill in a large matrix with predicted values}
\usage{
predict_all_parallel(cs, counts, verbose = T, ncores = 1)
}
\description{
Fill in a large matrix with predicted values
}
\keyword{internal}
