% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{predict_binned_matrices_irls}
\alias{predict_binned_matrices_irls}
\title{Perform binning of data}
\usage{
predict_binned_matrices_irls(cts, dispersion)
}
\arguments{
\item{cts}{cts}

\item{dispersion}{dispersion}

\item{verbose}{boolean.}
}
\value{
" "
}
\description{
fast IRLS and zero counts approximation
}
\examples{
" "
}
\keyword{internal}
