% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{prepare_difference_estimation}
\alias{prepare_difference_estimation}
\title{Build grouped difference matrix using normalization data if available}
\usage{
prepare_difference_estimation(cs, csg, resolution, ref, tol.val, nperf)
}
\description{
Build grouped difference matrix using normalization data if available
}
\keyword{internal}
