% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{prepare_first_signal_estimation}
\alias{prepare_first_signal_estimation}
\title{Prepare for concurrent signal estimation}
\usage{
prepare_first_signal_estimation(biases, names, base.res)
}
\description{
Prepare for concurrent signal estimation
}
\keyword{internal}
