% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{prepare_signal_estimation}
\alias{prepare_signal_estimation}
\title{Prepare grouped signal matrix and settings}
\usage{
prepare_signal_estimation(cs, csg, resolution, tol.val, nperf)
}
\description{
Prepare grouped signal matrix and settings
}
\keyword{internal}
