% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_and_prepare}
\alias{read_and_prepare}
\title{Wrapper function for the whole preprocessing}
\usage{
read_and_prepare(infile, outprefix, condition, replicate,
  enzyme = "HindIII", experiment = "Hi-C", name = paste(condition,
  enzyme, replicate), skip = 0L, nrows = -1L, locus = NULL,
  circularize = -1, dangling.L = c(0, 4), dangling.R = c(3, -1),
  maxlen = 600, read.len = 40, dmin = 2000, save.data = T)
}
\arguments{
\item{infile}{character. Path to TadBit .tsv file}

\item{outprefix}{character. Prefix to output intermediate files.}

\item{condition}{character. WT, KO, cell type etc.}

\item{replicate}{character. Replicate ID.}

\item{enzyme}{character. HindIII, DpnII etc.}

\item{experiment}{character. Hi-C Capture-C etc. For now it is Hi-C and
cannot be changed}

\item{name}{character. A name for the experiment. By default, it is
condition, enzyme and replicate}

\item{skip}{see \code{\link[data.table]{fread}}}

\item{nrows}{see \code{\link[data.table]{fread}}}

\item{locus}{if not NULL, a vector in the format c("chr3",123,456) to keep only
paired-end reads where both ends map in region chr3:123-456. Otherwise, read everything.}

\item{circularize}{integer. Set this to the size of the chromosome if it is
circular, otherwise leave as-is (default is -1)}

\item{dangling.L}{vectors of integers. Offset from cut site for
reads to be considered dangling ends (Left or Right respectively)}

\item{dangling.R}{vectors of integers. Offset from cut site for
reads to be considered dangling ends (Left or Right respectively)}

\item{maxlen}{positive integer.  Maximum admissible size of sonication 
fragment. Fragments above this threshold will be categorized as "other".}

\item{read.len}{integer vector. The length of a full read, used for dangling ends (temporary)}

\item{dmin}{numeric. Minimum disrance between cut sites to consider this as a
contact. Note that this setting will only be enforced upon merging, and the
returned object might contain contacts at a smaller distance.}

\item{save.data}{boolean. Whether to save a CSdata object that contains the
raw data}
}
\value{
A CSdata object
}
\description{
See \code{\link{read_tsv}}, \code{\link{categorize_by_new_type}} and 
\code{\link{prepare_for_sparse_cs_norm}} for further detail and arguments.
}
\examples{
" "
}
