% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromosome.R
\name{read_chromosome_binless}
\alias{read_chromosome_binless}
\title{Read matrices written by \code{\link{chromosome_binless}}}
\usage{
read_chromosome_binless(sample, sample2 = NA, zoom_region = "full",
  out_prefix = "chromosome_binless", read.raw = T, read.binless = T,
  read.signal = T, read.difference = T, read.signif = T)
}
\arguments{
\item{sample}{The name to use for the sample}

\item{sample2}{If difference matrices are to be loaded, the name to use for
the second sample}

\item{zoom_region}{If "full" (default), will load complete dataset. If a
vector c(start,end) is passed, will read only this portion.}

\item{out_prefix}{The output prefix used to store matrices.}

\item{read.raw}{Whether to load _raw_band.RData file(s). Default TRUE.}

\item{read.binless}{Whether to load _binless_band.RData file(s). Default
TRUE.}

\item{read.signal}{Whether to load _log10signal.RData file(s). Default TRUE.}

\item{read.difference}{Whether to load _log10diff.RData file (if applicable).
Default TRUE.}

\item{read.signif}{Whether to load _log10diff_signif.RData file (if
applicable). Default TRUE.}
}
\value{
mat data.table
}
\description{
This procedure produces a mat object amenable to \code{\link{plot_binless_matrix}}.
}
\examples{
" "
}
