% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{read_tsv}
\alias{read_tsv}
\title{Read TADBit tsv file and return (paired-end) reads as data.table}
\usage{
read_tsv(fname, nrows = -1L, skip = 0L, locus = NULL)
}
\arguments{
\item{fname}{The filename}

\item{nrows, skip}{see \code{\link[data.table]{fread}}}

\item{locus}{if not NULL, a vector in the format c("chr3",123,456) to keep only
paired-end reads where both ends map in region chr3:123-456. Otherwise, read everything.}
}
\value{
a data.table
}
\description{
Read TADBit tsv file and return (paired-end) reads as data.table
}
\examples{
" "
}
