% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{recover_normalization}
\alias{recover_normalization}
\title{Recover failed or aborted normalization}
\usage{
recover_normalization(prefix)
}
\arguments{
\item{prefix}{character. As provided to \code{\link{normalize_binless}} or
\code{\link{run_exact}}.}
}
\value{
A CSnorm object
}
\description{
Will look for files of the form prefix_lambdaxxx.RData and return the
simulation with the largest log-likelihood. Nothing else will be checked so
make sure these input files correspond to a run with the exact same
parameters up to the initial condition.
}
