% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_exact}
\alias{run_exact}
\title{Cut-site normalization (exact model)}
\usage{
run_exact(cs, bf_per_kb = 1, bf_per_decade = 20, lambdas = c(0.1, 1, 10),
  ncores = 1, iter = 1e+05, subsampling.pc = 100, init_alpha = 1e-07,
  prefix = NULL)
}
\arguments{
\item{cs}{CSnorm object as returned by \code{\link{merge_cs_norm_datasets}}}

\item{bf_per_kb}{positive numeric. Number of cubic spline basis functions per
kilobase, for genomic bias estimates. Small values make the optimization 
easy, but makes the genomic biases stiffer.}

\item{bf_per_decade}{positive numeric. Number of cubic spline basis functions
per distance decade (in bases), for diagonal decay. Default parameter 
should suffice.}

\item{lambdas}{positive numeric. Length scales to try out as initial
condition.}

\item{ncores}{positive integer (default 1). Number of cores to use.}

\item{iter}{positive integer. Maximum number of optimization steps for each leg.}

\item{subsampling.pc}{numeric. Percentage of the data used to do the calculations (default 100).}
}
\value{
A csnorm object
}
\description{
Will run the exact model of normalization (on one cpu for each lambda 
provided) and returns the most likely model and predicted quantities. Useful
for comparison purposes. If you don't know what to use, try 
\code{\link{normalize_binless}}
}
