% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{run_serial}
\alias{run_serial}
\title{Run exact model on a single cpu}
\usage{
run_serial(cs, init, bf_per_kb = 1, bf_per_decade = 20, iter = 1e+05,
  subsampling.pc = 100, init_alpha = 1e-07)
}
\arguments{
\item{cs}{CSnorm object as returned by \code{\link{merge_cs_norm_datasets}}}

\item{bf_per_kb}{positive numeric. Number of cubic spline basis functions per
kilobase, for genomic bias estimates. Small values make the optimization 
easy, but makes the genomic biases stiffer.}

\item{bf_per_decade}{positive numeric. Number of cubic spline basis functions
per distance decade (in bases), for diagonal decay. Default parameter 
should suffice.}

\item{iter}{positive integer. Maximum number of optimization steps for each leg.}

\item{subsampling.pc}{numeric. Percentage of the data used to do the calculations (default 100).}
}
\description{
Run exact model on a single cpu
}
\keyword{internal}
