% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{subsample_counts}
\alias{subsample_counts}
\title{Get the first ncounts/d of each of d datasets, including zeros}
\usage{
subsample_counts(cs, ncounts, dset = NA)
}
\description{
count is always a bit smaller because we censor those that are <dmin without adding more counts
}
\keyword{internal}
