% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{subsample_csnorm}
\alias{subsample_csnorm}
\title{Take a random subset of the reads of a given csnorm object}
\usage{
subsample_csnorm(cs, subsampling.pc = 100)
}
\arguments{
\item{subsampling.pc}{positive integer}

\item{csnorm}{a csnorm object}
}
\value{
The csnorm object with updated biases and counts slots
}
\description{
Take a random subset of the reads of a given csnorm object
}
\examples{
" "
}
