% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_dispersion.R
\name{theta.ml}
\alias{theta.ml}
\title{This function is adapted from MASS::theta.ml (GPL-3)
It allows to specify the initial value of theta, and removes superfluous checks}
\usage{
theta.ml(y, mu, weights, init.theta, limit = 25,
  eps = .Machine$double.eps^0.25, trace = FALSE)
}
\description{
This function is adapted from MASS::theta.ml (GPL-3)
It allows to specify the initial value of theta, and removes superfluous checks
}
\keyword{internal}
