% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{update_diagnostics}
\alias{update_diagnostics}
\title{update diagnostics data table or create it if not existing}
\usage{
update_diagnostics(cs, step, leg, runtime)
}
\description{
update diagnostics data table or create it if not existing
}
\keyword{internal}
