% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{zoom_csnorm}
\alias{zoom_csnorm}
\title{Restrict a CSnorm object to a given genomic portion.}
\usage{
zoom_csnorm(cs, begin, end)
}
\arguments{
\item{begin, end}{Start and end positions, refer to cs@biases[,pos]}
}
\value{
The csnorm object with updated biases and counts slots
}
\description{
Will remove any pre-existing normalization parameters
}
\examples{
" "
}
