% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{bin_all_datasets}
\alias{bin_all_datasets}
\title{Bin normalized datasets}
\usage{
bin_all_datasets(cs, resolution = cs@settings$base.res, ncores = 1,
  verbose = T)
}
\arguments{
\item{cs}{CSnorm object, normalized.}

\item{resolution}{integer. The desired resolution of the matrix.}

\item{ncores}{integer. The number of cores to parallelize the zeros calculation on.}

\item{verbose}{verbose}
}
\description{
Bin normalized datasets
}
\examples{
" "
}
