% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{bin_data}
\alias{bin_data}
\title{Bin a reads data.table into a matrix of a given resolution}
\usage{
bin_data(obj, resolution, b1 = NULL, b2 = NULL)
}
\arguments{
\item{obj}{a CSnorm object, or a data.table containing data to be binned}

\item{resolution}{the target resolution}

\item{b2, e2}{(optional) if provided, give extradiagonal portion}
}
\value{
a data.table representing the binned data. Will contain
  name,bin1,bin2 (bin coordinates), observed data and mean distance between
  these bins and nobs. If a CSnorm object is passed, will also compute the
  number of observables in this bin (i.e. 4x the number of cut site
  intersections). For the sake of speed, nobs is approximate along the
  diagonal as it assumes that all intersections have observables, which is
  not true since we discard all counts smaller than dmin.
}
\description{
Bin a reads data.table into a matrix of a given resolution
}
\examples{
" "
}
