% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binless.R
\docType{class}
\name{CSgroup-class}
\alias{CSgroup-class}
\title{Class for one dataset grouping at a given (base) resolution}
\description{
Class for one dataset grouping at a given (base) resolution
}
\section{Slots}{

\describe{
\item{\code{mat}}{data.table.}

\item{\code{interactions}}{list.}

\item{\code{resolution}}{in bases.}

\item{\code{group}}{}

\item{\code{cts}}{data.table. The counts and predicted means used in all calculations.}

\item{\code{names}}{}
}}

\keyword{internal}
