% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_binless.R
\name{detect_binless_differences}
\alias{detect_binless_differences}
\title{Binless detection of significant differences with a reference}
\usage{
detect_binless_differences(cs, ref, resolution = cs@settings$base.res,
  group = "all", ncores = 1, tol.val = cs@settings$tol, verbose = T,
  min.lambda2 = 0.1, fix.lambda1 = F, fix.lambda1.at = NA,
  fix.lambda2 = F, fix.lambda2.at = NA, nperf = 50)
}
\arguments{
\item{cs}{a normalized CSnorm object}

\item{ref}{The reference to compute the difference on, as obtained from cs@experiments}

\item{resolution}{The base resolution (default: the one used during normalization)}

\item{group}{Default "all", but also various groupings, see \code{\link{group_datasets}}}

\item{ncores}{How many cores to parallelize on (default: the one used during normalization)}

\item{tol.val}{Tolerance on computed quantities}

\item{verbose}{Whether to be verbose (default) or not}

\item{min.lambda2}{The minimum value of lambda2 when scanning for its optimum. Might
move a bit because of the CV+kSD criterion}

\item{fix.lambda1}{Default FALSE. if TRUE, sets lambda1
(resp. lambda2) to the provided value instead of optimizing it}

\item{fix.lambda1.at}{Default FALSE. if TRUE, sets lambda1
(resp. lambda2) to the provided value instead of optimizing it}

\item{fix.lambda2}{Default FALSE. if TRUE, sets lambda1
(resp. lambda2) to the provided value instead of optimizing it}

\item{fix.lambda2.at}{Default FALSE. if TRUE, sets lambda1
(resp. lambda2) to the provided value instead of optimizing it}

\item{nperf}{number of IRLS iterations (default 50)}
}
\value{
the binned matrix with additional information relating to these
  significant interactions
}
\description{
Binless detection of significant differences with a reference
}
