% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppDocs.R
\name{fast_binless}
\alias{fast_binless}
\title{Fast binless signal calculation}
\usage{
fast_binless(obs, nbins, alpha, lam2, lam1 = 0, nouter = 25, tol_val = 2e-1,
 bg_steps = 5, free_decay = 10000)
}
\arguments{
\item{obs}{DataFrame containing at least 4 named columns: name, bin1, bin2 and observed.
The first three must be convertible to integer vectors (like factors) and must range from
1 to N contiguously. For example, if there are 2 datasets, name should range from 1 to 2,
and similarly for the bin columns bin1 and bin2. observed is the number of reads that fall
at that coordinate. The provided matrices must be dense, i.e. report zero counts.}

\item{nbins}{unsigned integer referring to the number of bins in the provided matrix.}

\item{alpha}{the dispersion of the negative binomial}

\item{lam2}{numeric positive value for the fusion penalty, or a vector of
the same size as the number of datasets.}

\item{lam1}{numeric positive value (or vector) for the significance threshold (default is zero)}

\item{nouter}{unsigned The maximum number of iterations that should be performed (default 25)}

\item{tol_val}{double tolerance on the values for convergence and the fused lasso (default 2e-1)}

\item{bg_steps}{unsigned the maximum number of initial steps where no signal is fitted (default 5)}

\item{free_decay}{integer the distance in bases up to which the decay is not forced to decrease (default 10000)}

\item{compute_patchnos}{boolean whether to compute the patchno column or not (TRUE by default).
On very large datasets, this is slow.}
}
\description{
This function computes a fast approximation to the binless normalization of a given dataset
Most notably, it does not produce a statistically significant output, since the fusion and
significance thresholds must be set by the user. Other differences include a pointwise estimate
of the decay and genomic biases, and one decay and one genomic bias for all datasets. This fast 
model is meant for a quick overview of the data.
}
