% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CppDocs.R
\name{fast_binless_difference}
\alias{fast_binless_difference}
\title{Fast binless difference calculation}
\description{
Once normalized with \code{\link{fast_binless}}, differences with respect to a reference can be computed.
}
