% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso.R
\name{fused_lasso}
\alias{fused_lasso}
\title{run fused lasso on one dataset contained in matg, fusing 'valuehat' into
'value'}
\usage{
fused_lasso(csig, positive, fixed, constrained, verbose = T,
  fix.lambda1 = F, fix.lambda1.at = 0.1, fix.lambda2 = F,
  fix.lambda2.at = NA)
}
\arguments{
\item{csig}{the CSbsig or CSbdiff object to optimize}

\item{positive}{boolean. Constrain eCprime in order to force 'value' to be
positive?}

\item{fixed}{boolean. Set eCprime=0 throughout ?}

\item{constrained}{boolean. Constrain lambda1 so that any diagonal that contains
only one big patch be forced to have 'value'=0 ?}

\item{verbose}{boolean (default TRUE).

finds optimal lambda1, lambda2 and eC using BIC.}
}
\description{
run fused lasso on one dataset contained in matg, fusing 'valuehat' into
'value'
}
\keyword{internal}
