% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{gauss_genomic}
\alias{gauss_genomic}
\title{Single-cpu simplified fitting for iota and rho}
\usage{
gauss_genomic(cs, cts.common, verbose = T, update.eC = T)
}
\arguments{
\item{if}{a single value, use data for estimate of mu and that value as a
dispersion, otherwise it's a list with parameters to compute the mean from}
}
\description{
Single-cpu simplified fitting for iota and rho
}
\keyword{internal}
