% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{generate_genomic_biases}
\alias{generate_genomic_biases}
\title{Generate iota and rho genomic biases on evenly spaced points along the genome}
\usage{
generate_genomic_biases(cs, points_per_kb = 10)
}
\arguments{
\item{cs}{normalized CSnorm object}

\item{points_per_kb}{number of evaluation points per kb}
}
\description{
Generate iota and rho genomic biases on evenly spaced points along the genome
}
\keyword{internal}
