% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{get_cs_interaction_idx}
\alias{get_cs_interaction_idx}
\title{Fetch CSinter object from CSgroup object}
\usage{
get_cs_interaction_idx(csg, type, threshold = -1, ref = NULL,
  raise = T)
}
\arguments{
\item{type}{}

\item{threshold}{}

\item{ref}{}

\item{raise}{}
}
\description{
Fetch CSinter object from CSgroup object
}
\keyword{internal}
