% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize_binless}
\alias{normalize_binless}
\title{Binless normalization}
\usage{
normalize_binless(cs, restart = F, bf_per_kb = 50, bf_per_decade = 10,
  bins_per_bf = 10, base.res = 5000, ngibbs = 25, bg.steps = 5,
  iter = 100, verbose = T, init.dispersion = 0.1, min.lambda2 = 0.1,
  tol = 0.1, ncores = 1, fit.signal = T, fix.lambda1 = F,
  fix.lambda1.at = NA, fix.lambda2 = T, fix.lambda2.at = 2.5)
}
\arguments{
\item{cs}{CSnorm object as returned by \code{\link{merge_cs_norm_datasets}}}

\item{bf_per_kb}{positive numeric. Number of cubic spline basis functions per
kilobase, for genomic bias estimates. Small values make the optimization 
easy, but makes the genomic biases stiffer.}

\item{bf_per_decade}{positive numeric. Number of cubic spline basis functions
per distance decade (in bases), for diagonal decay. Default parameter 
should suffice.}

\item{bins_per_bf}{positive integer. Number of distance bins to split basis 
functions into. Must be sufficiently small so that the diagonal decay is 
approximately constant in that bin.}

\item{base.res}{base resolution to use for the fused lasso}

\item{ngibbs}{positive integer. Maximum number of iterations.}

\item{bg.steps}{The first steps are for fitting the background. Do at most bg.steps.}

\item{iter}{positive integer. Maximum number of optimization steps for each leg.}

\item{verbose}{Display progress if TRUE}

\item{init.dispersion}{positive numeric. Value of the dispersion to use initially.}

\item{min.lambda2}{The minimum value which lambda2 is allowed to take when searching its optimum. Too small values
can cause computational instability.}

\item{tol}{positive numeric (default 1e-1). Convergence tolerance on relative changes in the computed biases.}

\item{ncores}{positive integer (default 1). Number of cores to use.}

\item{fit.signal}{boolean. Set to FALSE only for diagnostics.}

\item{fix.lambda1}{whether to set lambda1 to a given value, or to estimate it. Default is estimate.}

\item{fix.lambda1.at}{if fix.lambda1==T, the approximate value where it is meant to be fixed. Might move a bit because
of the positivity and degeneracy constraints. Make sure this value is positive if lambda1 is fixed.}

\item{fix.lambda2}{whether to set lambda2 to a given value, or to estimate it. Default is fixed.}

\item{fix.lambda2.at}{if fix.lambda2==T, the approximate value where it is meant to be fixed. Make sure this value
is positive if lambda2 is fixed.}

\item{init}{boolean (default is FALSE). Whether to continue an existing
normalization or to start anew (default).}
}
\value{
A csnorm object
}
\description{
Binless normalization
}
\examples{

}
