% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.R
\name{predict_binned_matrices_irls}
\alias{predict_binned_matrices_irls}
\title{Perform binning of data}
\usage{
predict_binned_matrices_irls(cts, dispersion)
}
\arguments{
\item{verbose}{boolean.}
}
\description{
fast IRLS and zero counts approximation
}
\keyword{internal}
