% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gauss.R
\name{save_stripped}
\alias{save_stripped}
\title{Save normalized CS object, discarding everything but
input counts and final normalization parameters}
\usage{
save_stripped(cs, fname)
}
\arguments{
\item{cs}{a normalized CSnorm object}

\item{fname}{where to save it to}
}
\description{
Save normalized CS object, discarding everything but
input counts and final normalization parameters
}
